pushing TS_NEVER redefinition to max int64 and correcting erroneous recast of TIMESTAMP to int in sync core linkage functions for triplex_meter and eventgen

+++ powerflow/triplex_meter.cpp
@@ -1454,7 +1454,7 @@ static TIMESTAMP sync_triplex_meter_impl(OBJECT *obj, TIMESTAMP t0, PASSCONFIG p
 #ifndef __APPLE__
 extern "C" MODULE_API TIMESTAMP sync_triplex_meter(OBJECT *obj, TIMESTAMP t0, PASSCONFIG pass)
 {
-    return (int)sync_triplex_meter_impl(obj, t0, pass);
+    return sync_triplex_meter_impl(obj, t0, pass);
 }
 #else
 extern "C" MODULE_API TIMESTAMP sync_triplex_meter(OBJECT *obj, ...)
